TimeSlider Fixture
